Moved the Ceres loadFromROS functions into reusable functions in fuse_core
