+++ fuse_core/CMakeLists.txt
@@ -38,6 +38,7 @@ add_library(${PROJECT_NAME}
   src/async_motion_model.cpp
   src/async_publisher.cpp
   src/async_sensor_model.cpp
+  src/ceres_options.cpp
   src/constraint.cpp
   src/graph.cpp
   src/graph_deserializer.cpp
