+++ fuse_core/include/fuse_core/util.h
@@ -34,10 +34,14 @@
 #ifndef FUSE_CORE_UTIL_H
 #define FUSE_CORE_UTIL_H
 
+#include <ros/console.h>
+#include <ros/node_handle.h>
+
 #include <ceres/jet.h>
 #include <Eigen/Core>
 
 #include <cmath>
+#include <string>
 
 
 namespace fuse_core
@@ -148,6 +152,30 @@ Eigen::Matrix<T, 2, 2, Eigen::RowMajor> rotationMatrix2D(const T angle)
   Eigen::Matrix<T, 2, 2, Eigen::RowMajor> rotation;
   rotation << cos_angle, -sin_angle, sin_angle, cos_angle;
   return rotation;
+}
+
+
+/**
+ * @brief Helper function that loads strictly positive integral or floating point values from the parameter server
+ *
+ * @param[in] node_handle - The node handle used to load the parameter
+ * @param[in] parameter_name - The parameter name to load
+ * @param[in] default_value - A default value to use if the provided parameter name does not exist
+ * @return The loaded (or default) value
+ */
+template <typename T,
+          typename std::enable_if<std::is_integral<T>::value || std::is_floating_point<T>::value>::type* = nullptr>
+T getPositiveParam(const ros::NodeHandle& node_handle, const std::string& parameter_name, T default_value)
+{
+  T value;
+  node_handle.param(parameter_name, value, default_value);
+  if (value <= 0)
+  {
+    ROS_WARN_STREAM("The requested " << parameter_name << " is <= 0. Using the default value (" <<
+                    default_value << ") instead.");
+    value = default_value;
+  }
+  return value;
 }
 
 }  // namespace fuse_core
